backend/pyproject.toml
[project]
name = "macrosoft-backend"
version = "1.0.0"
description = "FastAPI backend for the Macrosoft Expense Tracker — LangGraph receipt auditor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # --- Web framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "httpx>=0.28.0",

    # --- LangChain / LangGraph ---
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openrouter>=0.1.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",

    # --- AI / Embeddings ---
    "google-genai>=1.68.0",

    # --- Data / Parsing ---
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "bs4>=0.0.2",

    # --- Utilities ---
    "python-dotenv>=1.2.2",
]
